+++ src/main/factories/signup/signup-validation.spec.ts
@@ -1,12 +1,8 @@
 import { makeSignUpValidation } from './signup-validation'
-import { ValidationComposite } from '../../../presentation/helpers/validators/validation-composite'
-import { RequiredFieldValidation } from '../../../presentation/helpers/validators/required-field-validation'
+import { ValidationComposite, CompareFieldValidation, RequiredFieldValidation, EmailValidation, PhoneValidation } from '../../../presentation/helpers/validators'
 import { Validation } from '../../../presentation/protocols/validation'
-import { CompareFieldValidation } from '../../../presentation/helpers/validators/compare-field-validation'
-import { EmailValidation } from '../../../presentation/helpers/validators/email-validation'
 import { EmailValidator } from '../../../presentation/protocols/email-validator'
 import { PhoneValidator } from '../../../presentation/protocols/phone-validator'
-import { PhoneValidation } from '../../../presentation/helpers/validators/phone-validation'
 
 jest.mock('../../../presentation/helpers/validators/validation-composite')
