+++ src/presentation/helpers/validators/index.ts
@@ -1,0 +1,5 @@
+export * from './compare-field-validation'
+export * from './email-validation'
+export * from './phone-validation'
+export * from './required-field-validation'
+export * from './validation-composite'
